pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "axiom-tfg"
version = "0.1.0"
description = "Deterministic physical task linter — feasibility gates for robotic task specs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Axiom Contributors"}]
dependencies = [
    "typer>=0.9,<1",
    "pydantic>=2.0,<3",
    "pyyaml>=6.0,<7",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "jinja2>=3.1,<4",
    "python-multipart>=0.0.9",
    "google-generativeai>=0.7,<1",
    "openai>=1.0,<2",
    "ikpy>=3.3,<4",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9",
    "httpx>=0.27,<1",
]

[project.scripts]
tfg = "axiom_tfg.cli:app"
axiom = "axiom_cli.app:app"

[tool.setuptools.packages.find]
include = ["axiom_tfg*", "axiom_server*", "axiom_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
